commit step 2

+++ backend/controllers/InvoiceController.php
@@ -3,18 +3,24 @@
 namespace backend\controllers;
 
 use backend\models\Artist;
+use backend\widgets\Str;
+use kartik\mpdf\Pdf;
+use PhpOffice\PhpSpreadsheet\Spreadsheet;
+use PhpOffice\PhpSpreadsheet\Writer\Xlsx;
 use Yii;
 use backend\models\Invoice;
 use backend\models\InvoiceSearch;
 use backend\models\InvoiceItems;
 use backend\models\InvoiceItemsSearch;
 use yii\base\InvalidConfigException;
+use yii\bootstrap\ActiveForm;
 use yii\data\SqlDataProvider;
 use yii\db\Exception;
 use yii\db\Query;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use yii\web\Response;
 
 /**
  * InvoiceController implements the CRUD actions for Invoice model.
@@ -60,16 +66,25 @@ class InvoiceController extends Controller
     public function actionView(int $id)
     {
         $searchModel = new InvoiceItemsSearch();
-        $dataProvider = $searchModel->search(['InvoiceItemsSearch' => ['invoice_id' => $id]]);
+        $query = Yii::$app->request->queryParams;
+        $query['InvoiceItemsSearch']['invoice_id'] = $id;
+
+        $dataProvider = $searchModel->search($query);
 
         $modelItems = new InvoiceItems();
 
+        $query = new \yii\db\Query();
+        $data = $query->from('invoice_items')
+            ->select('SUM(IF(artist_id != 0, `amount`, 0)) AS `total_artist`, SUM(amount) AS `total`')
+            ->where(['invoice_id' => $id])
+            ->one();
+
         return $this->render('view', [
             'model' => $this->findModel($id),
-            'items' => [
-                'dataProvider' => $dataProvider,
-            ],
+            'searchModel' => $searchModel,
+            'dataProvider' => $dataProvider,
             'modelItems' => $modelItems,
+            'total' => $data
         ]);
     }
 
@@ -91,7 +106,10 @@ class InvoiceController extends Controller
      */
     public function actionExport(int $id)
     {
-        $sql = "SELECT ss.artist_name, ss.track_name, ss.summ, ss.total 
+        $sql = "SELECT #ss.artist_name,
+                        #ss.track_name,
+                        ss.summ,
+                        ss.total 
                 FROM (
                 SELECT it.`artist_id`, a.name as artist_name, t.name as track_name, sum(it.amount) as summ, sum(DISTINCT(tt.am)) as total
                     FROM `invoice_items` it 
@@ -108,12 +126,13 @@ class InvoiceController extends Controller
                 ) as ss
                 WHERE ss.artist_id != 0 ORDER BY `ss`.`track_name` ASC";
 
+
+
         $data = Yii::$app->db->createCommand($sql)->queryAll();
 
         $file = \Yii::createObject([
             'class' => 'codemix\excelexport\ExcelFile',
             'sheets' => [
-
                 'Users' => [
                     'data' => $data,
                     'titles' => ['Artist', 'Track', 'Summ', 'Total'],
@@ -146,7 +165,8 @@ class InvoiceController extends Controller
 
 
     /**
-     * Displays a single Invoice model.
+     * Перерахунок суми інвойсу
+     *
      * @param integer $id
      * @return mixed
      * @throws NotFoundHttpException if the model cannot be found
@@ -175,6 +195,13 @@ class InvoiceController extends Controller
         return $this->redirect(['view', 'id' => $id]);
     }
 
+    /**
+     * Розрахунок інвойсу і оновлення депозівтів
+     *
+     * @param int $id
+     * @return \yii\web\Response
+     * @throws NotFoundHttpException
+     */
     public function actionCalculate(int $id)
     {
         $model = $this->findModel($id);
@@ -186,8 +213,47 @@ class InvoiceController extends Controller
             return $this->redirect(['view', 'id' => $id]);
         }
 
+        if ($model->invoice_type == 2) { // Виплата
+            $db = Yii::$app->db;
+
+            // записати останній інвойс і дату випалит в память акртисту
+            foreach ($model->getInvoiceItems()->all() as $item) {
+
+                $item->artist->saveBalance(3, $model->currency_id);
+
+                $db->createCommand()
+                    ->update('artist',
+                        [
+                            'date_last_payment' => date('Y-m-d', strtotime('2024-09-30')),
+                            'last_payment_invoice' => $model->invoice_id
+                        ],
+                        'id = :ID',
+                        [':ID' => $item->artist_id]
+                    )->execute();
+
+                $db->createCommand(
+                    "UPDATE `invoice_items` ii 
+                            INNER JOIN invoice i ON i.invoice_id = ii.invoice_id and i.invoice_type = 1 
+                         SET ii.`payment_invoice_id`= {$model->invoice_id} 
+                         WHERE ii.artist_id = {$item->artist_id} 
+                         AND ii.payment_invoice_id is null 
+                         AND i.currency_id = " .$model->currency_id
+                )->execute();
+            }
+
+            $db->createCommand(
+                "UPDATE aggregator_report_item ari 
+                            INNER JOIN aggregator_report ar ON ar.id = ari.report_id 
+                            INNER JOIN invoice i ON i.aggregator_report_id = ar.id and ar.report_status_id = 2 
+                            INNER JOIN invoice_items ii ON ii.invoice_id = i.invoice_id and i.invoice_type = 1 and ii.payment_invoice_id = {$model->invoice_id}
+                        SET ari.payment_invoice_id = {$model->invoice_id} 
+                        WHERE ari.payment_invoice_id is null 
+                            AND ari.isrc = ii.isrc"
+            )->execute();
+        }
+
         $model->calculate();
-        $model->invoice_status_id = 2;
+        $model->invoice_status_id = 2; // Розрахований
         $model->save();
 
         Artist::calculationDeposit();
@@ -203,6 +269,12 @@ class InvoiceController extends Controller
     public function actionCreate()
     {
         $model = new Invoice();
+
+        if (Yii::$app->request->isAjax && $model->load(Yii::$app->request->post())) {
+            Yii::$app->response->format = Response::FORMAT_JSON;
+
+            return ActiveForm::validate($model);
+        }
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
             return $this->redirect(['view', 'id' => $model->invoice_id]);
@@ -244,13 +316,43 @@ class InvoiceController extends Controller
     {
         $model = $this->findModel($id);
 
-            //if ($model->invoice_status_id != 2) {
+            if ($model->invoice_status_id != 2) {
                 $model->delete();
-          /// } else {
-           //     Yii::$app->session->setFlash('error', "Неможа видалити розрахований інвойст");
-          //  }
+           } else {
+              Yii::$app->session->setFlash('error', "Неможа видалити розрахований інвойст");
+            }
 
         return $this->redirect(['index']);
+    }
+
+    public function actionExportToExcel($id): void
+    {
+        $model = $this->findModel($id);
+
+        $filename = "/home/atpjwxlx/domains/blck.link/public_html/backend/web/report_artist_income_q{$model->quarter}_{$model->invoice_id}.xlsx";
+
+        if (file_exists($filename)) {
+            $this->redirect("/report_artist_income_q{$model->quarter}_{$model->invoice_id}.xlsx");
+        }
+
+        $spreadSheet = new Spreadsheet();
+        $workSheet = $spreadSheet->getActiveSheet();
+        $workSheet->setTitle('Дохід артистів по звіту');
+
+        $tempData = [];
+        $tempData[] = [
+            'Артист',
+            'Сума',
+            'Валюта'
+        ];
+
+        $tempData =array_merge($tempData,$model->getInvoiceReportDataGroupArtist());
+
+        $workSheet->fromArray($tempData, null, 'A1');
+        $writer = new Xlsx($spreadSheet);
+        $writer->save($filename);
+
+        $this->redirect("/report_artist_income_q{$model->quarter}_{$model->invoice_id}.xlsx");
     }
 
     /**
